First Half of Commenting Changes
This code was in a messy state after the first commit. This is halfway
towards making the code more readable and instructive to other editors
in anticipation of Winter Project Week 2017. More edits to come,
especially in the ROI.py script. Currently, the code is non-functional
because of renamed files.