Project Week - End of Wednesday - Bug/Usability Fixes
This is a morning update to the Github. Many fixes were added to remove
redundancy in this module's code, and it make it run more smoothly,
especially with the 3D rendering. Many thing are yet to be added for the
UI. Knowledge of Qt is relevant, but the up-front learning cost is
high.. It would be good for developers to have a wiki page in the future
for this sort of thing (TODO?)